Laying out code for the interpreter.
We have prepared the basic structure of the virtual machine that will
interpret the bytecode and layed out some basic instructions needed for
it to execute as well as some neccessities such as memory and registers that
will be utilized by the programs.

The design details are not important right now, that is a task for
another day and to be quite honest we are pretty flexible with the
design so there is nothing to stress about.