Adding unit tests for VM instructions.(CRLF->UNIX)
The Virtual Machine needs to be tested to make sure
it operates appropriately, we don't want some instructions
to do something they are not suppose to.

We Simply test the functionality of each instruction in a utility
tester class that simulates an usage of a Virtual Machine object as
and verfies that everything that had to happen, did happen!

All tests are yet to be implemented but the framework is layed, we
just have to do the tedious work of writing all the tests.

Converting source files to Unix formatting.

For some reason Visual Studio Code likes to convert/create source files
in MSDOS format CRLF instead of Unix. Because people here are actually
using Unix based platforms and text editors other than Microsoft's
Windows Notepad we don't want no carriage returns before ourf line feeds.