Generate parameters lists and parameter variant.
To pass parameters to an instructin without having to route them through
a switch statement we use an parameter list classes and than keep an
array of function callbacks that would implement the logic for each
instruction.

This would simplify the code that parses and executes the next
instruction and it wouldn't have to be generated. It would also simplify
implementing of the assembler and compilers in the future as well as
allow everything to be done during runtime without unnecessary code generation.